This contains updated data collection code for the server.  The new code aligns data from neural and behavioral sources.  Unfortunately, I (Matt) designed it to work after the fact, instead of in real time.  So, this code will mostly be thrown out (or heavily modified).
Also included in this update is server/datacollection/dbTest.zip.  This is a complete QT project that is used to create fake databases containing neural or behavioral data.  I used this to test the alignment code.  However, since we're running in real time, this is also not particularly useful, but is included for completeness.  The zip file will disappear in the next update.