This is a major update to the data collection and alignment for the server.  This update includes a neuraldatacollector object which polls a proxy server, and an alignmenttool object, which keeps track of the incloming alignment codes, and provides the fitting coefficitents for converting between timebases.
Additionally, a DebugCommandHandler was created.  The debug command was created to allow the developers to execute test code on the server.  It is expected that this handler will have lots of different targets for test purposes.  Te command handler should probably not be included in the release version of the code.