I modified Director so that when the server crashes, it backs out to the "looking for server" mode (instead of crashing).
I also added a "reset" function to the STate object.  This returns the visual stimuli in the state to their original condition.  In other words, if there is a square in the top-right corner that gets relocated by a script when State::run() is called, the next time we run the state, the square will still be in the top-right corner.