Support for fail safe system operation is about halfway complete.  At this point, the server keeps a filebacked database with a list of opened sessions.  If the server crashes and comes back online, it can reload any session using this list and the sessions database.  This reloading occurs automatically when a component instance connects to the server with a session that is in this list of opened sessions.  When sessions are cleanly ended, they are deleted from the opened sessions database.  They are also deleted if there has been no activity on the session in the database for over 1 day.  In this way, at least from the servers view, if it gets disconnected or crashes, so long as the components don't crash, when it gets reconnected or starts back up it can pick up where it left off.
Everything up to that point is now complete.  The incomplete part is the various components side of things.  The director needs to be able to continue operating even if it looses its connection to the server.  It also has to maintain a record of all data that wasn't sent to the server and saved to disk on the server so that that data can be resent.  This is what I will be working on next.