A lot has changed.  The State Machine Editor is now basically functional, and it is possible to build experiments with it.  Undo and Redo are enabled and it is possible to transition from StateMachineEditor to Text Editor back and forth with contents updating and undo/redo enabled for both.  I have changed the XML syntax somewhat such that States now contain all visual and audio elements along with their control elements.  This makes the parameter system more logical.  At this point, every script has access to the scriptable members (ie. Parameters or Visual Elements) that are declared inside of it or any of its ancestors.
I also made the whole StateMachine tree much more object oriented which made the StateMachineEditor code much cleaner.  The system now has ResultContainers, ScriptableContainers, and MachineContainers that are the ancestors of objects that contain results, scriptable items, and transitions respectively.  Everything that is visible in the main StateMachineEditor window is now also a UIEnabled object, which means that it has a UIInfo child that contains (currently only) visual block position information.  This object may contain more information in the future (ie. user notes sbout a particular object, highlight colors, size, etc.).  Both Parameters and Visual Elements are also now Scriptables.  Using all of these ancestor classes I have moved a whole ton of code that was being repeated in multiple locations and coded in multiple ways into single spots where they can be tailored to meet the needs of the application and the UI.  

I added notes to the UI window that shows which scriptables are available for a selected item with script capabilities.  This currently is showing too many functions, but I will fix that shortly.  

I have removed FlowElements, PredicateExpressions and CompoundExpressions from the solution.  Although these were pretty well built up and contained lots of code, their functionality is easily emulated in slightly adjusted ScriptElements (they now return either a true or false result).  Given that users are going to have to script a little anyway, it is a whole lot easier to script out the complicated boolean functions of the parameters than it is to create them with the UI as was the idea in Flow Elements.  I have not deleted the Flow Element files from the latest version on the Trunk; however, in case I think of some reason that they should be kept around.  

Everything is built now according to the idea that each MachineContainer (ie. StateMachine or State) runs until it hits a result.  When a result is hit, operation continues in that MachineContainer's parent.  The Task ends when a result in its topmost MachineContainer is hit. 

Finally, due to all these changes I had to disable some of the Unit Tests because they no longer reflected the code that is running.  Hopefully, when we get to the test stage we can fix some of these.