Fixed bug with new entry and exit target scripts in TargetController and ChoiceController.  Their scripts were getting run even when the experiment was running in slave mode which caused them to overwrite some of the values received from the server.  These scripts are now ignored in slave mode (as all scripts should be).
Also added code to the D3DVisualTarget and PixmapVisualTargets to report when their latest FirstPhosphor time was.  This time is used to mark the frame time in state.cpp and soon I will update the system to mark all property changed with the first phosphor times as well.  WARNING:  AT THIS STAGE, THE PIXMAP VISUAL TARGET DOES NOT REPORT FIRST PHOSPHOR CORRECTLY.  TO BE SURE OF CORRECT FIRST PHOSPHOR FUNCTIONALITY, THE D3DVISUALTARGET SHOULD BE USED.  (ie. Don't put the -pixmap tag in the call to run the director).