Fixed bug that was creating dataids for data units when they are serialized instead of when they are created.  This means that ordering SQLITE queries by dataid didn't always return things in the order that they happened.
Got rid of the value field on the timerparameter so that its not constantly updating the server with useless data and is simply used to get the time since it was reset.  This change should probably happen in general with parameters since they can have no effect on the monkey's experience.  It needs some careful consideration, but possibly, the fields that frequently change on parameters should be removed so that the server doesn't get unnecessary data.  We would still keep these kinds of fields around in visual elements and possibly a new kind of Info Element.