Changed analysis viewer to the bare bones version of the long term analysis plan.  Now, the operator fills in the blanks of a statements as follows: When element/property at path: ___________ changes/starts/ends, print value of property at path: ______________.  "Append time since"/"Start at first time" element/property at path _________________ changed/started/ended.
The various options are filtered into drop down boxes according to the chosen element/property type.  This gives the operator a very general way to select out values from the experiment and see when they changed relative to a given time.  The long term plan is to allow multiple items in the "when" section at which the value will be printed.  Mutiple values will also be allowed in the "print" section which will be separated on output by commas.  The "append time" section at the end will just be another "print" option.  Then we will follow this with a "start at" section.  Also there will be an option of printing whether a state was started/ended between some other start/end selections in this section.  None of this is set in stone though and it may change.  Also, keep in mind that the robust functionality of this concept depends on our adding transitions to initial elements within the state machine, so we'll have to do that too.

In this commit, I've also included some small updates to the various experiments.  Using the current iteration of the analysisviewer, we will end up missing the last trial in the experiment's trial info in analysis because of the element at which TrialInfo.updateValue() was called in these experiments.  We have fixed this issue in the experiments here, and once we have a more robust analysis viewer as described above, we will be able to get all the TrialInfo even from experiments run prior to these changes.