Minor change to plexon plugin.  Now units are numbered starting at zero.  Before they were numbered by letter 4a, 4b, 4c, etc., which just turned into an ascii value on the server anyway.
Note: After some research, while the lfp values were correct for the data we looked at, the waveform data for spikes seemed more complex.  We checked the values, and they were correct but off by a scale factor, and in fact they are simply the values returned by the plexon A/D converter.  Since we're starting experimentation with a tdt system anyway, we're going to leave this alone for a while, but at some point we should figure out a way to automatically write the correct voltage values to the picto server and not just the waveform A/D read outs like we're doing now.  

As far as TDT goes, at this stage I'm just seeing zeros in the waveform readings, so it looks like I'll need to do some more research...