Major Update:
After discovering the lack of support for multiple Nidaq tasks working with the same port, I changed the signalChannel architecture such that now signal channels are now associated with an input port.  The InputPort is a single port that may be associated with multiple signal channels each of which registers the port channels that it will be looking at and their sample rate.  The InputPort automatically calculates the maximum sample rate that it can use for the port as a whole while meeting the demands of the individual channel's sample rates, and by using this greatest common devisor sample rate, it can use internal counters to return the samples to the signal channel at the rates that the signal channel requested.

During the course of making these changes I discovered a signal channel alignment issue.  Previously, all signal channels were started at the beginning of a task and considered to have perfect timing throughout for the purpose of timestamp calculations.  This turned out to not be correct since about a ms of misalignment error arose for every 100 sec during on test.  To fix this, the InputPort code is now autoaligning eye data.  It samples the underlying hardware at 4 times the necessary level, then compares calculated sample times with picto timestamps during runtime and throws out samples that are calculated to have times greater than 1 sample period beyond the picto timestamp and duplicates samples that are calculated to have times greater than 1 sample period less than the picto timestamp.  Since alignment is run every frame, the clock is good enough that we won't misalign by more than one sample period per frame, and we are sampling at least 4 samples for every requested data sample, we can be assured that these corrections won't lower our resolution.