Major Update to Analysis System:
Analysis is now defined by XML files just like Experiments.  An AnalysisDefinition is the top level object with the top level XML tag and defines the analysis for an experiment.  The AnalysisDefinition object contains at least one AnalysisPeriod, which is the unit period over which analysis will be performed.  In a typical experiment, an analysis period will be a single trial.  AnalysisPeriods are defined by a StartTrigger and optionally an EndTrigger.  Pre and post window time buffers can be added as well.  AnalysisPeriods contain at least one trigger.  There are multiple types of triggers, SpikeTrigger, TransitionTrigger, PropertyTrigger, etc.  Start Triggers and End Triggers from the Analysis Period are used by that object to determine the time window for each analysis period.  All other AnalysisTriggers contain at least on AnalysisDataSource.  Each time an AnalysisTrigger is triggered, it reads in all of its child AnalysisDataSources and puts them in a new row of a script array.  There are multiple types of AnalysisDataSources as well.  PropertyDataSource, LFPDataSource, etc.  At the end of each AnalysisPeriod, a user defined script is run that can access the arrays that were built up during the current period in each of the AnalysisPeriod's AnalysisTriggers.  The script can then use functions made available from AnalysisTools that are also children of the AnalysisPeriod to do things with this information.  Currently there is one analysis tool which is a FileOutput tool.  Scripts can call the writeLine function in this object to write a line of data out to a file, that is stored in a temporary folder beneath the application run directory.

Soon, the AnalysisWidget of the workstation will have an option to save the contents of the temporary output directory to another location.  In the future, the plan is to add other types of AnalysisTools as well, like plotter, task progress plots (like on the right side of the window in Orion), etc, that will can be run on incoming data in the remote viewer.