THIS IS AN IMPORTANT COMMIT! THIS IS THE CODE BEING USED IN THE PICTO SYSTEM AS CHRIS STARTS RECORDING NEURAL DATA FOR THE FIRST TIME!
Changes:
Added the capability to put AnalysisTools in the AnalysisDefinition for the purpose of writing out data on a "per run" basis.  This meant that we could update the AmbiguityAnalysisFull.xml analysis definition to write out the Run Notes from each run with its output data.
I also fixed the TaskRunViewer UI so that users that are not authorized in a session don't have any buttons available to modify Run info.
One important note is that currently AnalysisPeriods are unable to access the scriptable values from their parent AnalysisDefinitions.  This breaks the logic that is used in Experiment definitions and so we should probably enable this.  This would also allow users to do useful things like adding data calculated during analysis into a run level notes file.