Updated Analysis system to support multiple start and end triggers per period.
In Analysis, a period was defined as running from a StartTrigger to the next StartTrigger or from a StartTrigger to the following EndTrigger and only one of each could be defined.  In Chris' experiment, he was running Analysis periods from a StartTrigger to the next StartTrigger and since in the case of the last Trial there was no final StartTrigger to close the period, the last Trial couldn't be analyzed.
We had the option to fix this in a number of ways.  We could have added an option to use the end of a Task Run as an EndTrigger or we could have allowed for multiple EndTrigger definitions.  
In the interest of generality, we went with option 2 and modified the analysis system to allow for multiple EndTriggers as well as multiple StartTriggers.  The first StartTrigger to appear defines the beginning of a period, and the first EndTrigger to appear after that StartPeriod defines the end of the period.  Subsequently, the next StartTrigger to appear defines the beginning of the next period, ignoring any StartTriggers that may have occurred between last period's StartTrigger and EndTrigger. etc.