Upgraded the Qt Infrastructure to Qt 4.8.1.
Note that it will be necessary to go through the "Building a Development Environment" instructions starting at the "Building Qt" section in order for this latest code to work on existing development machines.

The Qt upgrade was part of an effort to speed up the Experimental Analysis which was limited by the speed of Qt Scripting.  Apparently, between 4.5 and 4.8 there was a drastic Scripting speed increase, which is why it seemed like a good idea to go ahead with the upgrade.  I have fixed the "Building a Development Environment" instructions on the Wiki to reflect the new Qt version.
I also made some changes to the Analysis system to speed things up when I noticed that one of the major speed bottlenecks was the Script->C++ function calls.  Now all of the Analysis Data Source data is loaded into script arrays which translates into a significant speed increase for Analyses that request Data Source data large numbers of times in loops (ie. eye and lfp data).
This version hasn't been extensively tested yet (and since we changed the underlying Qt installation, it should be tested), but I will do some testing before delivering it to Chris.