This is the version of Picto that we are using to start Hyojung with the Matching Pennies experiment on the PictoBox.
Latest Changes:
File output now only opens its underlying file for Text if the first function called on it is "writeText" otherwise it opens it for binary data.  This is important because writing binary 0x0A when the file is open for text has the result of appending a 0x0D since 0x0A is a line break and 0x0D is an automatically added carriage return.

Also, fixed issue in FileOutput where binary float was being written as 64 bit double (the default in 1t 6.4 and up).

Update to the Matching Pennies experiment.