Moved Slave Experiment system to an event driven non-blocking setup to significantly reduce code complexity.  Now, the SlaveExperimentDriver continuously pushes updates from the StateUpdater to the Experiment and runs a single frame of experiment for each new frame pushed from the StateUpdater.
Updated RemoteViewer to work with new slave experiment system.

I would love to make the director itself event driven rather than forcing it to all run in a singal monolithic thread as we are doing now.  New hardware would probably be sufficient to allow us to maintain the frame rate.  The only issue is detection of the first phosphor time.  This will require some research.