Upgraded to Qt 5.0 and Visual Studio 2010.
A lot of files changed here.  The majority of the file changes are minor (ie. QString::getAscii() is deprectated and changed to .toLatin1()); however, there was significant effort in streamlining the process of creating a development environment as well as updating the various qt .pro and .pri files to support the new Qt version.  
Since Qt has now started including the QtPropertyBrowser in its source tree, we have started using that instead of downloading our own copy.  We also eliminated the zip libraries since we don't use them anyway and updated the Proxy plugins to support the new Qt plugin system.  
The directory structure of the Picto.sln solution is now somewhat different.  This is due to the fact that our old vcProjFilters application that set up the solution directory structure no longer works with VS2010; however, QT offers a "CONFIG -= flat" configuration option that automatically creates a reasonable VS directory structure for us based on the file system structure and we will be using this instead.
I have updated our vcSlnDependencies code/application to work with VS2010 and we are still using that as part of the GenerateBuildFiles process.
I have not yet updated the Wiki with the new instructions for setting up the development environment.  I am doing that next.