Added colored bars for scripts in green bars on Elements in StateEditViewer. Added tooltip which shows underlying script when hovering over a particular colored bar.
Changed PropertyBrowser such that script properties only show up when selecting the start bar now. Also added large green start bar for window asset on left side of canvas.

Changed Script properties in the start bar such that are minimizable and start expanded if they contain text. Emptying a script of text (or leaving only whitespace) has the effect of making the marker in the startbar for that script disappear. The algorithm for detecting an empty script in the start bar is simplified by the fact that the scriptTextEditor erases any whitespace when the user focuses out of it so long as there is no other non-whitespace text in the editor.