Added capability to open experiments from session files. Updated StateEditViewer to show script elements on red result bars underneath asset items. Updated Run Order of Operations such that now: 1. State Entered. 2. EntryScript called. 3. Begin Loop     a. Do Control logic and break if done.     b. Run frame script     c. Present frame.     d. Run AnalysisFrameScript. 4. Run exit script.
The only affect this had on pre-existing experiments was that I needed to change the upgrade system such that target entry/exit scripts were put into the state exit script as well.  This is just in case target exit causes the state to exit, meaning anything looking for target exit in the state's frame script would never get called.