Rough Sketch of Analog Input framework
Extensively modified SignalChannel class.  Now SignalChannel is a psuedo-Abstract class, and all the scaling functionality (including that which restricted it to being used with data packets that just had an X and Y channel) have been moved to the DoubletSignalChannel subclass. GenericInput is a subclass of SignalChannel meant for handling Analog Inputs.  Analog Inputs use a separate data store for serialization, InputDataUnit.