New FixationController added to address issues with existing TargetController
The FixationController operates over a fixed period, and uses reacquisition to determine when a break in fixation is sufficient to be considered a failure.  Otherwise, the fixation time continues to progress.

Also added the ability to mark Warnings for Assets, which are displayed in Red in the AssetElementBox.

Removed an unnecessary file.
Updated documentation for changes.
Continued to make small documentation improvements.
Updated the version number.